requirements.txt
django>=4.2.0,<5.0.0
pillow>=9.5.0
numpy>=1.24.0
opencv-python-headless>=4.7.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.5
